requirements.txt
flask
google-generativeai
flask-cors